ignore paths support2
